requirements.txt
Flask
python-telegram-bot
schedule
rtorrent-python